[FIX] microsoft_calendar: fix occurrences sync
In Outlook, like any event, a recurrence is identified by an `id` which is
specific to the Outlook user calendar, and by an `iCalUId` which is the same
among all Outlook calendars.

Each event of an Outlook recurrence (called `occurrence`) uses a `seriesMasterId`
attribute to refer to the Outlook recurrence. Unfortunately, this `seriesMasterId`
attribute refer to the `id` which is specific to the user calendar and not to the
universal `iCalUId`. That means, when a user B is syncing his Odoo calendar and gets
a recurrence created by a user A, to be able to sync occurrences with Odoo event, we need
to first match the occurrence `seriesMasterId` attribute with the Odoo recurrence through
the universal ID of this recurrence.

Part-of: odoo/odoo#82863